mcp/plugins/cursor_clone/requirements.txt
# Requirements for Cursor-AI Clone
# Core dependencies
pyyaml>=6.0
structlog>=24.0.0

# Web panel
fastapi>=0.100.0
uvicorn>=0.23.0

# HTTP client for OpenAI API
httpx>=0.26.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: LLM backends
# llama-cpp-python>=0.2.0  # Uncomment if using llama.cpp
# transformers>=4.30.0     # Uncomment if using HuggingFace

